added scraper code
