+++ server/server.py
@@ -1,5 +1,4 @@
 from flask import Flask
-from scraper import get_internships
 
 app = Flask(__name__)
